pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aie"
version = "0.1.0"
description = "Atmospheric Intelligence Engine: reference implementation for urban air quality forecasting."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "TBD", email = "ddaydota@gmail.com" }]
keywords = [
    "air quality",
    "deep learning",
    "forecasting",
    "atmospheric science",
    "uncertainty quantification",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Atmospheric Science",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "numpy>=1.26,<2.0",
    "pandas>=2.1",
    "scipy>=1.11",
    "scikit-learn>=1.4",
    "xgboost>=2.0",
    "torch>=2.2",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "openpyxl>=3.1",
    "pyarrow>=15.0",
    "rich>=13.7",
    "properscoring>=0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "black>=24.3",
    "mypy>=1.8",
    "pre-commit>=3.6",
    "jupytext>=1.16",
]

[project.urls]
Homepage = "https://github.com/research-AI-studies/atmospheric-intelligence-engine"
Repository = "https://github.com/research-AI-studies/atmospheric-intelligence-engine"
Issues = "https://github.com/research-AI-studies/atmospheric-intelligence-engine/issues"

[project.scripts]
aie = "aie.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
aie = ["py.typed"]

# ---------------------------------------------------------------------------
# Tooling configuration
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = ["notebooks", "data"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM", "RUF"]
ignore = ["E501"]  # handled by formatter

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
no_implicit_optional = true
pretty = true

[tool.pytest.ini_options]
addopts = "-ra --strict-markers"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]
